Let sysui integration tests run in presubmit.
TEST_MAPPING doesn't allow arbitrary intersection of annotations, so
this changes how we label sysui @Scenario tests slightly:

Before:

@Ignore / @FlakyTest: doesn't run at all.
@Staging: runs in staged-postsubmit only.
else: runs in postsubmit (and staged-postsubmit).

After:

@Ignore / @FlakyTest: doesn't run at all.
@Staging: runs in staged-postsubmit only.
@Postsubmit: runs in postsubmit (and staged-postsubmit) only.
else: runs in presubmit (and postsubmit and staged-postsubmit).

The upshot of this is that new tests will need to be marked @Staging or
@Postsubmit until they're known not to be flaky, or presubmit will break
a lot.

Bug: 165922296
Test: atest --test-mapping frameworks/base/packages/SystemUI:presubmit
Change-Id: If5aa903ed19bbc072f347c736251644cedcfce1a

Former-commit-id: 49887b270c58518fe9606a12c058cee1448349a3