Pass interaction type to FalsingManager.isFalse
This ensures that the falsing manager always is testing against the
correct interaction type when it is asked. Prior to this change, we
could end up in states where a user was swiping down (opening the
quick settings) but the falsing manager believed the user was trying
to unlock the phone.

Fixes: 160967364
Test: atest SystemUITests && manual
Change-Id: I176f54a768622dec3e758a7f01ed8aec26223648
Merged-In: I176f54a768622dec3e758a7f01ed8aec26223648

Former-commit-id: bf88e9faf1e9d892dee24896c1bb5a944b355e0a