Stop providing duplicate info to NAS
Since the NAS is a notification listener, it has access to
notifications in its current user and enabled managed profiles.
However, unlike other listeners, it also has an instance running in
the managed profile.

This adds extra filtering on 'current user' before sending information
to the NAS so the NAS doesn't have to filter itself.

Test: CTS tests for listeners and assistants
Test: atest
Fixes: 154938387
Change-Id: If8dee2ac6d6ce71fdaf6f43bd240b361e8828686

Former-commit-id: 6cd9d5eea1734942c378b840dc276e4c5bd228e7