Move SystemServicesModule to Global Scope.
Rename SystemServicesModule to FrameworkServicesModule to make the
name less ambiguous.

Include the GlobalModule in all the various GobalRootComponents.

Include the WMModule in all the various GlobalRootComponents (it
is currently empty so its exclusion was not breaking anything, but
it should soon!)

LocalBluetoothManager has to be moved out of the module as it requires
a handler for instantiation.

Bug: 162923491
Test: manual
Change-Id: Id8498b661759ce0846ee904398dfcfe20ee744e9

Former-commit-id: ff4809911c81a63534f9c4c1f0dafdc0820ae8cf