Cached list of pre-created users on UserManagerService.
PermissionManagerService calls UserManagerService to get the list
of user ids multiple times on boot (O(1000) times), so these calls
must be cached. PMS used to call UMS.getUserIds(), which didn't
include the list of pre-created users, so it was recently changed
to call UMS.getUsers() and calculate the ids, but such non-cached
call increase the boot time on some devices in almost 50ms.

This change fixes the regression by caching the user ids of the
pre-created users as well.

Test: manual verification
Bug: 165940683

Change-Id: Ib4068cce7d4cea56e0937099508469e49e3412e2

Former-commit-id: d4407fd5304cd71f6a42d04f59c59d943efc427a