Append error code with error message
Before ag/12117351, we used to create an extra apk session for
verification of apks in session, and listened for result via an intent
receiver. If verification for the extra session failed, the error
message would reach StagingManager via dispatchSessionFinished call
which appended the error code with error message.

On ag/12117351, we stopped creating the extra session and attached the
error message on staged session directly. This broke
CtsInstallHostTestCases#DowngradeTest which expects
to find "INSTALL_FAILED_VERSION_DOWNGRADE" error code string in staged
session error message.

Bug: 159336213
Test: atest DowngradeTest
Change-Id: I1c169e01f028dface1e5152d62522f839e34ac76

Former-commit-id: d4324c4ced81fa91d507d08047b73c6aed46c060