Run appsFilter.systemReady() last
This change moves the appsFilter systemReady call to the end of the PMS
systemReady method to avoid it acquiring the package lock on a
background thread as the lock is required by several other blocks in the
method.

Before: MakePackageMangerServiceReady timing was 300-500ms, most of that
time spent setting up preferred activities as it waited for mLock to
become available
After: MakePackageManagerSerivceReady timing is ~130ms.

Test: atest AppEnumerationTests
Fixes: 162347084
Fixes: 163613646
Change-Id: Iac2948fb2ad66e45161627f637f039a5f4faf88c

Former-commit-id: 112a64a6dd94903f56272bc5c0d11d32f7b4e2ed