Swap seekbar vert padding to keep height fixed
Reduce the vertical padding of the seekbar when it's enabled by 2dp to
account for the extra height in the track. That way the height of the
SeekBar view is always 33dp.

This has a nice side effect of properly setting the height of the track
when resuming.

Bug: 160242262
Test: manual
Change-Id: I9b6888f2a6bf5fb118075dea3f849618350ec18b
(cherry picked from commit a6cd4ac4797ac73a49839f02bca88d74f7bfbdb4 [formerly b84e6ba27f61885a4c5282f19fbbf238bcaeaa91])

Former-commit-id: 6f3c88ab863360e2ca05988f06b6aa3fde3fa14b