Magnification button applies fade-in and fade-out animation
We use runnable task to handle the UI animation to prevent the case
that "Disable animation" settings disables all the animation showing.
The default showing duration of magnification button is 3 secs and
the fade-in & fade-out animation durations are both 300ms.
The showing step: 1.fade-in > 2.show magnification button > 3.fade-out
And the demo video is attached in the bug

Bug: 168867867
Test: atest MagnificationModeSwitchTest
Change-Id: If2d9a93987555b947e36e09b5345aaf588c02927

Former-commit-id: 5d7e9fabf63511cc1b59609ccd0c276a7752a434