Configure Bluetooth profiles displayed in settings
Two profiles were displayed on automotive targets at the wrong times,
MAP was displayed for all devices, PBAP was displayed only when
enabled.  This updates those two profiles to appear in the profiles only
when the service is enabled locally and the remote device supports the
other half of the connection.

Bug: 173818456
Bug: 109783106
Test: pair with bluetooth devices that support (and don't support) MAP
and PBAP, and verify settings slider appears when appropriate
specifically when disabled.

Change-Id: I62b8778517730076554ba996cee80862ce72539b
Merged-In: I62b8778517730076554ba996cee80862ce72539b

Former-commit-id: e4dc456d4b0fb474ca3821967b0913c820a713f7