Remove remaining insets system UI flags in DisplayPolicy
This patch is to remove all the remaining insets hidding and immersive
flags, including:
- SYSTEM_UI_FLAG_HIDE_NAVIGATION
- SYSTEM_UI_FLAG_FULLSCREEN
- SYSTEM_UI_FLAG_IMMERSIVE
- SYSTEM_UI_FLAG_IMMERSIVE_STICKY

Translucent and transparent flags are removed and the new appearance
flags are used instead. The removed flags are:
- View.STATUS_BAR_TRANSPARENT
- View.STATUS_BAR_TRANSLUCENT

BarController functions got removed, and the necessary channels for
informing System UI the insets state change got moved to DisplayPolicy
instead.

Flags are replaced with the new API's. Some necessary modifications are
made to support the new API.

layoutNavigationBar and layoutStatusBar will always return false,
removed dead code for it.

Test: go/wm-smoke
Test: atest DisplayPolicyTests
Test: atest DisplayPolicyLayoutTests
Test: atest DisplayPolicyInsetsTests
Bug: 155413635
Change-Id: I74404f2abe5df521e097f7d7b2bb7cc93ea9328c

Former-commit-id: 9913411b547ae5bf92233499a7a12e8faef6bee7