Make TV volume icons outlined instead of filled
This change adds new svns for the TV icons and adds a translateX tag to
move them left/right so that they align.

Here we also add a config for whether to have different icons for volume
"off" and volume "mute". The new icon ic_volume_media_off is used to
signify when the volume has been disabled (e.g. by a dedicated remote control
button press). This is different from ic_volume_media_mute, which signifies that
the volume has been set to 0. In the "off" position, the volume is also set
to 0, but the volume dialog remembers the previous setting, so when +/-
is pressed, this is relative to the previous setting.

By default ic_volume_media_off is the same as ic_volume_media_mute. Only on
drawable-television/ do we set a dedicated icon for ic_volume_media_off.

Bug: 167732325
Test: m && flash && verify icons have been changed
Change-Id: I8d6a6c2b971974894e5062d652bf905f38bf8e83

Former-commit-id: 5e6ce51fb90afb5722ffd0ac292cd87d69df919b