Make task visible if activity launched behind
Activities were unable to be started while dream started and
keyguard locked because the started activities were not on the
top-most focused tasks (it was the dream task on top).

Even the activities were set as launched behind, the visibilities
of the belonging tasks were still invisible and aborted the
launches.

Making the task visible if the top activity is launched behind
and also starting activities on all visible tasks for attached
application.

Bug: 168326273
Test: making phone call on Auto while Dream started on Phone
Test: atest DreamManagerServiceTests
Test: atest ActivityRecordTests
Change-Id: I199435bcfc7593e6c58488cd6de7c2289d5d0e20

Former-commit-id: 24d2622f748d1245c61cf9cafa22a4a6768bed5c