System API annotation and API .txt changes
New APIs are being exposed for use by SetUp Wizard, which requires
System APIs.

The APIs enable the calling app to determine the capabilities for /
config of the current user that affect time zone detection and then
alter them.

This commit also contains changes to the Shell package needed for the
CTS tests being added to test the new APIs. See the associated CTS
commit for details.

The primary usecase:

The rules for who can do what and when for time zone detection are
becoming more complicated, and device and user capabilities will be
affected by compile time constants, device config (resources) and device
hardware. By exposing this API, the complex logic can be encoded in one
place (the system server), rather than duplicating them in multiple
client apps.

These APIs are also intended for use by SettingsUI (and the various form
factor versions of that), but settings apps do not require them to be
marked with @SystemApi.

Bug: 159891384
Test: atest CtsTimeTestCases
Change-Id: I1eb6a68a74217106f81cb9525ad9370c805eecf2

Former-commit-id: 9ec32ea53490bbdcdf01be28a6e2fa56e69d2e68