Revert "Create wrapper around Shader native implementation"
This reverts commit 4a614493d0723c4fd76c812d4f339b8f80f14e66 [formerly fc42a99ea5712883c9872d34a523d972c26c9d6f].

Because we are moving away from a wrapper class to delegate
between SkImageFilter and SkShader and instead adding
RenderEffect on RenderNode, this wrapper implementation
is no longer necessary. There are some behavioral
differences between SkShader and SkImageFilter that
would lead to additional complexity in conversions
between these objects and potential bugs.
To simplify this, RenderEffect will be added
directly to RenderNode to support better
caching of rasterization layers that is necessary
for various visual effects like blur.

Test: N/A
Change-Id: I65dc9f8bd26aed8761af13491ae3f6d583208377

Former-commit-id: 5bed1f5332c54924368d9692c6b104fd5373576a