Build stub libraries with smaller deps
The stub libraries had a dependency to framework-internal-util which is
a collection of java static libraries whose source code is
auto-genmerated (e.g. via protoc, aidl, etc.)

The library was added as a dependency because some Java files that are
fed into the metalava for the stub code generation have references to
some of the auto-generated symbols. However, this is more than
necessary. Furthermore, that setup will create a circular dependency
when framework-internal-utils (or its transitive dependencies) have a
dependency to the stub libraries due to 'sdk_version' setting.

Fixing the problem by replacing the dependency to
framework-internal-util with a few number of libraries.

Bug: 169618843
Test: m
Change-Id: I30c953ed220e84ab295548845fff0b26e3c060b6

Former-commit-id: c0f8a28e54ba5d5d2691e7eaf319c7ed7b387d0c