WM: Prepare BLASTSync to flip main BLAST flag.
We will flip BLAST on in multiple stages. First the main
BLASTBufferQueue adapter, and in a second stage BLASTSyncEngine
buffer passing. We protect the usage of this buffer passing behind
a compile time flag so we can flip the main flag while continuing
development on this. While we are here, we also take care to
look at this flag from when processing the MainWindowSizeTransaction
if BLASTSync is disabled we need to use deferred transactions to make
this work, but if BLASTSync is enabled, deferred transactions will
actually sabotage us, preventing the buffer from latching.

Bug: 152346200
Test: Existing tests pass
Change-Id: I8a5d700c4b2e0da814ed12da3fb8521cba8849da

Former-commit-id: 2333ecb654eb794d8c1aadb5f1221a8cd6ec282d