Expose grantImplicitAccess in IPackageManager
This change exposes the method to grant implicit visibility access via
IPackageManager and as a hidden API in PackageManager. This variant of
the method takes a recipient UID and the authority that it should see
and limits access to only the contacts provider on device.

Bug: 158688602
Test: PackageManagerTests
Change-Id: I0050593e4aa734af1a69a40a60746f7cf0ea72df
Merged-In: I0050593e4aa734af1a69a40a60746f7cf0ea72df
Merged-In: I2c80513a5dffa7ad92963e376ea85217f127f39a

Former-commit-id: 48afc4448c649b0c3af0f8a97532eb36000d0b61