Fix mic indicator crash when animation gets cancelled
Take into account the fact that when ValueAnimator#cancel() is called
the listener receives the onAnimationEnd() as well as the
onAnimationCancel(). The fact that we did not destinguish between the
scenarios when the animation is fully completed and when the animation
was cancelled led to the crash in the AudioRecordingDisclosureBar. This
CL fixes the crash by handling these two cases differently.

Bug: 166578426
Test: make, flash
Test: adb shell appops start <package> 27
Test: adb shell appops stop <package> 27
Change-Id: I19489e3db18aeabfac8bbd3df0a04b7d85f5a697

Former-commit-id: 38f85b72c6cc2c3a884d1040f994c51bc2688fba