Add last_released checking for non-updatable API
This makes the build fail when APIs are removed from the (non-updatable)
framework.jar, and e.g. into a module. This is in general not safe, so
is a reasonable default. When it needs overriding (when e.g. creating
now modules), incompatibilities.txt can be used for that.

Bug: 170105267
Test: m checkapi
Change-Id: If608be48276f0a7f76afd714656267ed379ce062
Merged-In: If608be48276f0a7f76afd714656267ed379ce062

Former-commit-id: 3258a479d3eef3af5da6aeec87cdf74405cc1cb6