Use hasBadge instead of isManagedProfile
To determine whether a user has a badge, the
correct call is now hasBadge, rather than isManagedProfile.

Currently, the only user that actually can have a badge
is a managed profile, so this change is a no-op. In the future,
if further profile support is expanded, then hasBadge will
be correct.

Bug: 170249807
Test: Treehugger (this cl is a no-op)
Change-Id: I9144c4dc519936dfc5a2aeae0eb0bd91bea89db2

Former-commit-id: 82194b8fe0e51e733663c99695d588fff9c811d9