Revert^2 "Use same display for activities in same process."
The error was a bad git merge that removed an import.
The fix was just to redo the commit again so the import is available.

Original Commit Message:

Use same display for activities in same process.

Use WindowProcessController to calculate display to hold activity.
Use the display with the top most activity within the same application.
If no top activity is available use the default display
Add additional parameter to task params modifier to include request.
Handle launching on same display if using PendingIntent.

Bug: 152958720
Test: atest WmTests:WindowProcessControllerTests
      atest WmTests:TaskLauncherParamsModifierTests

Change-Id: If96cdc7207ec9dba6663eb327350e7148605f157

Former-commit-id: c32262be8bcf5f874d4c73f14f811e126ceca568