Add logs for screenshot errors
Currently, we post a notification but do not log anything for some
classes of screenshot error (specifically, if the bitmap is null
or if the service connection is disrupted). This means that even
if a user reports that an error ocurred, we have no way to tell
which error it was or what the cause might have been.

Bug: 170310033
Fix: 170310033
Test: manual (deliberately caused the relevant errors to verify
that the logs were observed)

Change-Id: I18a32655e0f353e8befa55e030434dda0c4a2071

Former-commit-id: 4a0fedc2e3f8eaa5ae0e5830c3b0f2136eceed5a