Add AIDL connections for SipDelegateManager
Implement AIDL backing for SipDelegateConnection,
DelegateConnectionStateCallback, and
DelegateConnectionMessageCallback as well as all
Parcelables related to those callbacks.

Implement SipDelegateConnection with an AIDL wrapper
implementation, which trampolines callback events
from the remote process to the local callback interface
as well as wrap the SipDelegateConnection AIDL so the
local process can send messages to the remote SipDelegate.

The next CL will be the ImsService AIDL implementation for
SipTransport/SipDelegate APIs.

Test: atest FrameworksTelephonyTests
Change-Id: Ic48cd0f80058d59362c63cf6381304437a81afe0

Former-commit-id: bf9c71073384e5266a4b8aae47a2220502758dbc