Fix windowing mode for non-resizable tasks
This contains 4 connected fixes for making the Dream activity
non-resizable and fullscreen:

1.  Non-resizable tasks should not be allowed to be in a multi-window
mode.

2. Task#mForceResizableActivities should only be applied to "standard"
activities (eg, Home and Dream should not be forced resizable).

3. Dreams should be set to non-resizable to ensure the activity behind
is not visible, as asserted by the DreamManagerServiceTests CTS test.

4. To ensure the non-resizable dream activity is launched fullscreen,
always create a new one-level task for the Dream - avoid creating an
additional child task as it is not needed, there is only ever one
DreamActivity.

Bug: 168775742
Test: atest CtsWindowManagerDeviceTestCases:DreamManagerServiceTests
Change-Id: Iffe57061cac2044699da78d84d64df0dddfb3a26

Former-commit-id: e1fb953f8571e01723ef2ce6f1b975fde5e564ff