Third refactor for BasePermission.
Make mPermissionInfo non-null and remove duplicate fields, to make
sure we have a single source of truth.

The previous nullness of mPermissionInfo is now equivalently
represented by the boolean field mReconciled.

mPendingPermissionInfo is also removed, which results in a slight
change of behavior but I believe it's okay looking at the logic.

Bug: 158736025
Test: presubmit
Change-Id: Ic346a7c3e2811326da3748368af32153b921476d

Former-commit-id: 7b6dc8af9009961ef2defab70c67767ca4601e56