Perf: optimize getCertHash
SHA-1 was the only option in the GP SEAC standard GPD_SPE_013 since 2014.
GPD_SPE_068 said SHA-256 shall be used in 2017, which also became the de
factor standard in the industry. In the most cases (if not all), it's
not necessary to check SHA-1 at all. The SHA-1 code should be removed in
the near future as long as it's confirmed not necessary.

Bug: 168624720
Test: Confirmed only SHA-256 is calculated. The dogfood phone works with
master local build.

Change-Id: Id9f09cf200f02d72daf230b2dbf789c9f985e2b0
Signed-off-by: Li Li <dualli@google.com>

Former-commit-id: feee500cfa16976565db74b8e8263f54a2ae3842