Only rely on ActivityConfigurationChangeItem to update activity config
This removes the legacy path that updates or set the pending activity
config when process config is changed. Since the handling of activity
config is separated, the legacy path from process config change almost
becomes no-op because the activity override config will dominate the
result of config difference.

This removal may also avoid the config change happened on unexpected
timing if the process registered to an activity as a config listener.
And the condition (in collectComponentCallbacks) for deciding whether
to deliver config to activity is not consistent with task manager
service on server side.

Bug: 171389102
Test: CtsWindowManagerDeviceTestCases:AppConfigurationTests
Change-Id: Ia4f9a9cc3538f5f419bdc9d125e2a930074e2f86

Former-commit-id: 06ccd09f912d3409397ddbefe70c6de0e2df1fa6