Seperate out DisplayWindowSettings persistence from retreival/policy logic. (1/n)
Previously DisplayWindowSettings was responsible for both settings lookup
and persistence as well as display setting policy. This CL breaks up
this responsibility leaving DisplayWindowSettings reponsible for
settings policy and migrating settings retreival/persistence to
DisplayWindowSettingsProvider.

It also changes the way the display window settings files are
maintained. Before, the display window settings file provided in vendor/
was copied on boot into data/system if there did not already exist a
file in data/system. The file in data/system was then overwritten when
new display settings were set (for ex, adb shell wm density). Now, the
settings file in vendor/ is no longer copied into data/system and the
settings values that are returned are a merged result of the two settings
files (with the file in data/system overwriting/overlaying the settings
provided in vendor/). This allows supporting the functionality of
disabling the settings provided in vendor/ while keeping the display
settings override functionality (adb shell wm density) intact.

Test: atest WmTests:DisplayWindowSettingsTests
Test: atest WmTests:DisplayWindowSettingsProviderTests
Bug: 168808369

Change-Id: Icab8e7d1f346cf991298056d098325e3b766a841

Former-commit-id: 759f1370c690f0ca026f22af53222105f814ae48