Add a QS tile for Reduce Bright Colors
The flow will be the same as that of our other accessibility feature,
Color Inversion.
1) RBC is in the stock list of tiles users can move to the panel
2) (TODO)  The first time the feature is turned on, it gets moved to
the panel

Long-pressing the tile navigates to the RBC subpage, where users
can adjust the intensity slider.

TODOs: call into CDS for config availability, get a tile icon,
update strings, set a callback for adding tile, so the first time
the feature is turned on, it gets moved to the panel

Test: long-pressing opens the RBC settings page, atest
ReduceBrightColorsTileTest
Bug: b/128465252

Change-Id: I26959fedf71298830fa5325307f555c8e2bc3901

Former-commit-id: 9ee2301f7115987f2e4309e53a95cb1d413029df