Do not transition from gesture detection state after gesture cancelation.
The agreed-upon behavior has always been that after gesture cancelation, touch explorer would exist in essentially a dead state.
It should not be possible to execute a two-finger passthrough gesture for example after doing an invalid gesture.
This is likely why we see GESTURE_PASSTHROUGH in cases where we expect GESTURE_UNKNOWN.
Fix: 171421786
Test: manual
Change-Id: Idce921c5278dea174ec5dad5cecf046f12ab113d

Change-Id: If9aeb611417bba08edd9f5a11f3d107c0b477bb8

Former-commit-id: 1cd4bd9ccf491c54fd03ae3d35ea763b5ea2630e