Migrate QuotaController to DeviceConfig.
This migrates flags inside QuotaController to DeviceConfig.
DeviceConfig is the recommended way and using it means
that we only operate/parse the set of flags that changed instead of
every single flag whenever one of them changes.

This uses the JobScheduler DeviceConfig namespace instead of creating a
wholly separate namespace just for the controller.

Bug: 124466289
Test: atest CtsJobSchedulerTestCases
Test: atest FrameworksMockingServicesTests:ConnectivityControllerTest
Test: atest FrameworksMockingServicesTests:JobSchedulerServiceTest
Test: atest FrameworksMockingServicesTests:JobStatusTest
Test: atest FrameworksMockingServicesTests:QuotaControllerTest
Test: atest FrameworksServicesTests:JobStoreTest
Change-Id: Id397dccc6d9c40bac1a9837addcab514f509a721

Former-commit-id: 12fce9938eb06e6b994b35990f1974876ea58b1b