Add null checking for Log.wtf CrashInfo
This is a hidden API that is only used inside RuntimeInit. When a caller
claims to be the system, ActivityManagerService can't necessarily tell
if they're being honest and assumes they are. The request is put on a
background handler to avoid locks, and the app is not supposed to be
allowed to crash.

We weren't checking well-formedness of the other arguments. A
NullPointerException on the background handler can take down the
entire system. So we should check the crashInfo for nullness before
it's used.

Test: atest FrameworksCoreTests
Bug: 171963789
Change-Id: I12341acef9b36e8d2a1af9952f9af93a5210a67a

Former-commit-id: b7c28f9429a4366402024bb493f0c16cb8f506bb