Implement Phase 4 of per-display power states
Remove the concept of mGlobalDisplayState and instead track the state of
each display individually. The internal methods of DisplayManagerService
have been updated to allow for displays to update individually, however
for this phase there is temporary code which still makes every display
update when the default display state is updated.

Bug: 138328918
Test: manual
Change-Id: I3b3d8df1d0fc0cd7934474a3664a69bde585504f

Former-commit-id: 8346e4070f2c4f07e1fee48878a1ffea013ffd24