Update the bouncer to be able to move to either side of a wide screen.
This is currently enabled by settings
persist.sysui.keyguard.one_handed_keyguard_enabled to true. If enabled,
this will cause "one-handed able" security modes to be measured with
half of the width of the SecurityViewContainer. Tapping either side of
the screen will then move the keyguard by changing the margins.

There's some flaws here, but it's a start. There's a frame of the
bouncer being on the wrong side of the screen if the keyguard is on the
right, and it doesn't respond to opening/closing the device properly yet.

Bug: b/170858298
Change-Id: Ib997b9b3cb4ca7a069c805c7a8f3a79434eebd4c
Test: Incoming :)

Former-commit-id: 6c5bfcb40fdcf2499098daad37577e7ead67699a