Make SnapshotStartingWindow and its child surface trusted
Touches were being blocked by go/untrusted-touches feature when the user
started interacting with the app early on while this window was being
displayed.

Original intent of the window is to let touches pass through (ie it sets
FLAG_NOT_TOUCHABLE), the window is posted by the system and the content
controlled by the system (a snapshot of the app).

So, we make the window trusted and allow touches to pass through.

Bug: 171683699
Bug: 158002302
Test: atest WmTests
Test: Open app for 2nd time and tap on the screen quickly, verify no
      toast about untrusted touch.
Test: Verify in input dumpsys that window is trusted.
Change-Id: I018ebf9e2e27a81c553749a05a2b343a25144d9d

Former-commit-id: a03b86f3d74289c67fdc84eceb431ad419f3ce6e