Increase StartInputHistory entry size
We have been receiving an increasing number of reports that say the
device went to a weird state where the keyboard stopped showing up.

One of the challenges when diagnosing this types of issues is that the
reporter often keeps using the device after triggering the bugreport
and the most important logs in StartInputHistory were already gone
when InputMethodManagerService#dump() got called.  Ideally IMMS should
be able to suppress logging while a bugreport is being taken, but
let's simply increase the entry size for now.

Bug: 171637033
Bug: 171792138
Bug: 171597353
Test: adb shell dumpsys input_method
Change-Id: I10450c7b78908a8c16949fa94f76fc76156a3bc2

Former-commit-id: 2f77697ea8498cd61322f3541c05ad90e3b1e7c9